The file cursor can be made completely private.
It could be passed as an argument, as needed. That would be a better
design, actually.
