+++ link-grammar/dict-file/dictionary.c
@@ -24,7 +24,6 @@
 #include "dict-common/file-utils.h"
 #include "dict-common/idiom.h"
 #include "dict-common/regex-morph.h"
-#include "dict-file/dict-file.h"
 #include "dict-ram/dict-ram.h"
 #include "post-process/pp_knowledge.h"
 #include "read-dialect.h"
@@ -229,15 +228,8 @@ dictionary_six_str(const char * lang,
 	                          /*align*/false, /*exact*/false);
 
 	/* Read dictionary from the input string. */
-	File_Dictionary fdict = malloc(sizeof(struct File_Dictionary_s));
-	dict->file_data = fdict;
-
-	fdict->input = input;
-	fdict->pin = fdict->input;
-	if (!read_dictionary(dict))
-	{
+	if (!read_dictionary(dict, input))
 		goto failure;
-	}
 
 	if (NULL == affix_name)
 	{
