+++ link-grammar/dict-file/read-dict.h
@@ -19,7 +19,7 @@ Dictionary dictionary_six(const char *lang, const char *dict_name,
                           const char *pp_name, const char *cons_name,
                           const char *affix_name, const char *regex_name);
 Dictionary dictionary_create_from_file(const char *lang);
-bool read_dictionary(Dictionary dict);
+bool read_dictionary(Dictionary dict, const char *input);
 void dict_error2(Dictionary dict, const char *s, const char *s2);
 
 void insert_list(Dictionary dict, Dict_node * p, int l);
